Task 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -24,7 +24,7 @@ blogsRouter.post('/', async (request, response) => {
     return response.status(400).json({ error: 'UserId missing or not valid' })
   }
 
-  console.log("user, from router", user)
+  //console.log("user, from router", user)
 
   const blog = new Blog({
     title: body.title,
@@ -46,8 +46,26 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).json({error: 'only the creator can delete the blog'})
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
